feature: [Interceptor] Change id section placement
