+++ apps/ai-dial-admin/src/components/Interceptors/View/Properties/Properties.tsx
